pyproject.toml
[project]
name = "rymflux"
version = "0.1.0"
description = "A CLI tool to stream and enjoy audiobooks and podcasts from the web."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]",
    "beautifulsoup4",
    "PyYAML",
    "typer",
    "rich",
    "questionary",
    "nest-asyncio>=1.6.0",
    "pyqt6>=6.9.1",
    "prompt-toolkit>=3.0.51",
]

[project.scripts]
rymflux = "rymflux.cli.main:start"
rymflux-gui = "rymflux.gui.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
